requirements.txt
# ── Core inference ───────────────────────────────────────────
sentence-transformers>=2.7.0
numpy>=1.24.0

# ── Data handling ────────────────────────────────────────────
pandas>=2.0.0

# ── Web interface ────────────────────────────────────────────
streamlit>=1.35.0

# ── Environment variables ────────────────────────────────────
python-dotenv>=1.0.0

# ── ML backend (CPU-only wheel; remove +cpu suffix for GPU) ──
torch>=2.1.0
